resources: move deployment construction into a file/function
For simplification of logic an later reuse.
