+++ internal/resources/smbservice.go
@@ -17,12 +17,9 @@ package resources
 
 import (
 	"context"
-	"strings"
 
 	appsv1 "k8s.io/api/apps/v1"
-	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -105,57 +102,8 @@ func (m *SmbServiceManager) deploymentForSmbService(s *sambaoperatorv1alpha1.Smb
 	// stop using hard coded parameters.
 	cfg := conf.Get()
 	// labels - do I need them?
-	labels := labelsForSmbService(s.Name)
-	var size int32 = 1
-
-	dep := &appsv1.Deployment{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      s.Name,
-			Namespace: ns,
-			Labels:    labels,
-		},
-		Spec: appsv1.DeploymentSpec{
-			Replicas: &size,
-			Selector: &metav1.LabelSelector{
-				MatchLabels: labels,
-			},
-			Template: corev1.PodTemplateSpec{
-				ObjectMeta: metav1.ObjectMeta{
-					Labels: labels,
-				},
-				Spec: buildPodSpec(cfg, s.Spec.PvcName),
-			},
-		},
-	}
-
+	dep := buildDeployment(cfg, s, ns)
 	// set the smbservice instance as the owner and controller
 	controllerutil.SetControllerReference(s, dep, m.scheme)
 	return dep
-}
-
-// labelsForSmbService returns the labels for selecting the resources
-// belonging to the given smbservice CR name.
-func labelsForSmbService(name string) map[string]string {
-	return map[string]string{
-		// top level labes
-		"app": "smbservice",
-		// k8s namespaced labels
-		// See: https://kubernetes.io/docs/concepts/overview/working-with-objects/common-labels/
-		"app.kubernetes.io/name":       "samba",
-		"app.kubernetes.io/instance":   labelValue("samba", name),
-		"app.kubernetes.io/component":  "smbd",
-		"app.kubernetes.io/part-of":    "samba",
-		"app.kubernetes.io/managed-by": "samba-operator",
-		// our namespaced labels
-		"samba-operator.samba.org/smbservice": labelValue(name),
-		"samba-operator.samba.org/share":      "share",
-	}
-}
-
-func labelValue(s ...string) string {
-	out := strings.Join(s, "-")
-	if len(out) > 63 {
-		out = out[:63]
-	}
-	return out
 }
